Busisness Layer/ MongoDB
defining models for movies_API Non relational database with MongoDB
Creating Endpoints and testing them via CRUD method / Mongoose schema for models. Using import and export modules.

The struggle here was connecting the server and understanding the issues :
-  slow  internet
- slow laptop / broke
- installing and executing the modules correctly
- command prompt vs powershell vs git bash
-request in postman put vs post vs get vs delete
pay attention to syntax
easy to clone a repo from GitHub but How to connect and merge local repo to existing repo on GitHub without duplicating on and on